feat : PostCSS : Run PurgeCSS only on production build. feat : App.svelte : Add base for layout search w/ tailwindcss.

+++ postcss.config.js
@@ -1,13 +1,14 @@
 const tailwind = require('tailwindcss');
 const autoprefixer = require('autoprefixer');
 const PurgeCSS = require('@fullhuman/postcss-purgecss');
+const prod = process.env.NODE_ENV === 'production';
 
 module.exports = {
     plugins: [,
         tailwind,
         autoprefixer,
-        new PurgeCSS({
-            content: ['./src/**/*.js', './src/00/*.svelte'],
+        prod && new PurgeCSS({
+            content: ['./src/**/*.js', './src/**/*.svelte'],
             css: ['./src/*.css'],
             defaultExtractor: content => content.match(/[A-Za-z0-9-_:/]+/g) || []
         }),
